// app/types/config-stie.ts
export interface ConfigSite {
  id: number;

  // Thông tin chung
  email: string;
  hotline: string;
  description: string;
  infoContact: string;
  infoFooter: string;
  image: string;
  favicon: string;
  googleMap: string;
  googleAnalytics: string;
  place: string;
  aboutImage: string;
  aboutText: string;
  aboutUrl: string;

  // Mạng xã hội
  facebook: string;
  zalo: string;
  instagram: string;
  linkedin: string;
  tiktok: string;
  twitter: string;
  x: string;
  youtube: string;
  pinterest: string;
  liveChat: string;
}